fix(review): complete calendar extraction to fix dependency inversion
Move the entire iCal subsystem (types, parser, calendar-api) and tests
from features/assignments/ to common/services/calendar/, eliminating
the common/ → features/ dependency inversion flagged by both reviewers.

- Move ical/types.ts, ical/parser.ts, ical/index.ts to common/services/calendar/ical/
- Move calendar-api.ts to common/services/calendar/
- Move associated test files alongside their modules
- Update all imports across assignments hooks, auth services, and pages
- Remove now-unused backward-compat re-export shims
- Fix import ordering lint warnings
