+++ packages/web/src/features/assignments/hooks/useCalendarAssociationFilter.test.ts
@@ -1,7 +1,7 @@
 import { renderHook, act } from '@testing-library/react'
 import { describe, it, expect, beforeEach } from 'vitest'
 
-import type { CalendarAssignment } from '@/features/assignments/api/calendar-api'
+import type { CalendarAssignment } from '@/common/services/calendar/calendar-api'
 import { useCalendarFilterStore } from '@/common/stores/calendar-filter'
 
 import { useCalendarAssociationFilter, ALL_ASSOCIATIONS } from './useCalendarAssociationFilter'
