+++ packages/web/src/features/assignments/hooks/useCalendarAssociationFilter.ts
@@ -12,8 +12,8 @@ import { useMemo, useCallback, useEffect } from 'react'
 
 import { useShallow } from 'zustand/react/shallow'
 
+import type { CalendarAssignment } from '@/common/services/calendar/calendar-api'
 import { useCalendarFilterStore, ALL_ASSOCIATIONS } from '@/common/stores/calendar-filter'
-import type { CalendarAssignment } from '@/features/assignments/api/calendar-api'
 
 // Re-export for backward compatibility
 export { ALL_ASSOCIATIONS } from '@/common/stores/calendar-filter'
